Create A Solution which will allow you to calculation the length of longest substring in a given
This piece of code, written in python, allows you to find out the length of the longest substring in a given string. The solution is written in python programming language and is implemented using classes. If I'm wrong or made a mistake, then your suggestions are very much appreciated. Hopefully this will help you in some way.
Kind Regards,
Muhammad Hashir Baig.